CSI raw block volume support

+++ csi/server/plugin/opensds/node.go
@@ -16,6 +16,7 @@ package opensds
 
 import (
 	"fmt"
+	"os"
 	"os/exec"
 	"strings"
 
@@ -54,17 +55,37 @@ func getVolumeAndAttachment(volumeId string, attachmentId string) (*model.Volume
 }
 
 // mountDeviceAndUpdateAttachment Mount device and then update attachment
-func mountDeviceAndUpdateAttachment(device string, mountpoint string, key string, mountFlags []string, needUpdateAtc bool, attachment *model.VolumeAttachmentSpec) error {
+func mountDeviceAndUpdateAttachment(device string, mountpoint string, key string, mountFlags []string, needUpdateAtc bool, attachment *model.VolumeAttachmentSpec, block *csi.VolumeCapability_BlockVolume) error {
 	var err error
 
-	if len(mountFlags) > 0 {
-		_, err = exec.Command("mount", "-o", strings.Join(mountFlags, ","), device, mountpoint).CombinedOutput()
-	} else {
-		_, err = exec.Command("mount", device, mountpoint).CombinedOutput()
-	}
+	if nil == block {
+		if len(mountFlags) > 0 {
+			_, err = exec.Command("mount", "-o", strings.Join(mountFlags, ","), device, mountpoint).CombinedOutput()
+		} else {
+			_, err = exec.Command("mount", device, mountpoint).CombinedOutput()
+		}
 
-	if nil != err {
-		return status.Error(codes.Aborted, fmt.Sprintf("failed to mount: %v", err.Error()))
+		if nil != err {
+			return status.Error(codes.Aborted, fmt.Sprintf("failed to mount: %v", err.Error()))
+		}
+	} else {
+		_, err = os.Lstat(mountpoint)
+
+		if err != nil && os.IsNotExist(err) {
+			glog.V(5).Infof("Mountpoint=%v is not exist", mountpoint)
+		} else {
+			glog.Errorf("Mountpoint=%v already exists!", mountpoint)
+			_, err := exec.Command("rm", "-rf", mountpoint).CombinedOutput()
+
+			if nil != err {
+				return err
+			}
+		}
+
+		err = os.Symlink(device, mountpoint)
+		if err != nil {
+			return err
+		}
 	}
 
 	// update volume Attachmentment
@@ -242,49 +263,66 @@ func (p *Plugin) NodeStageVolume(
 
 	// Check if it is: "Volume published but is incompatible"
 	mnt := req.VolumeCapability.GetMount()
-	mountFlags := mnt.MountFlags
-	_, err = exec.Command("findmnt", device, mountpoint).CombinedOutput()
-	glog.V(5).Infof("findmnt err: %v \n", err)
+	block := req.VolumeCapability.GetBlock()
+	glog.V(5).Infof("VolumeCapability Mount=%+v, Block=%+v\n", mnt, block)
 
-	if nil == err {
-		if len(mountFlags) > 0 {
-			_, err := exec.Command("findmnt", "-o", strings.Join(mountFlags, ","), device, mountpoint).CombinedOutput()
-			if nil != err {
-				return nil, status.Error(codes.Aborted, "Volume published but is incompatible")
+	if (nil != mnt) && (nil != block) {
+		return nil, status.Error(codes.InvalidArgument, "volumeMode cannot be both Block and Filesystem")
+	}
+
+	if nil == vol.Metadata {
+		vol.Metadata = make(map[string]string)
+	}
+
+	var mountFlags []string
+	if nil == block {
+		vol.Metadata[KCSIVolumeMode] = "Filesystem"
+		mountFlags = mnt.MountFlags
+		_, err = exec.Command("findmnt", device, mountpoint).CombinedOutput()
+		glog.V(5).Infof("findmnt err: %v \n", err)
+
+		if nil == err {
+			if len(mountFlags) > 0 {
+				_, err := exec.Command("findmnt", "-o", strings.Join(mountFlags, ","), device, mountpoint).CombinedOutput()
+				if nil != err {
+					return nil, status.Error(codes.Aborted, "Volume published but is incompatible")
+				}
+			}
+
+			return &csi.NodeStageVolumeResponse{}, nil
+		}
+
+		// Format
+		curFSType := connector.GetFSType(attachment.Mountpoint)
+		hopeFSType := DefFSType
+		if "" != mnt.FsType {
+			hopeFSType = mnt.FsType
+		}
+
+		if "" == curFSType {
+			_, err := exec.Command("mkfs", "-t", hopeFSType, "-F", device).CombinedOutput()
+			if err != nil {
+				return nil, status.Error(codes.Aborted, fmt.Sprintf("failed to mkfs: %v", err.Error()))
+			}
+		} else {
+			if "" != mnt.FsType {
+				if mnt.FsType != curFSType {
+					glog.Errorf("Volume formatted but is incompatible, %v != %v!", mnt.FsType, curFSType)
+					return nil, status.Error(codes.Aborted, "Volume formatted but is incompatible")
+				}
 			}
 		}
 
-		return &csi.NodeStageVolumeResponse{}, nil
-	}
-
-	// Format
-	curFSType := connector.GetFSType(attachment.Mountpoint)
-	hopeFSType := DefFSType
-	if "" != mnt.FsType {
-		hopeFSType = mnt.FsType
-	}
-
-	if "" == curFSType {
-		_, err := exec.Command("mkfs", "-t", hopeFSType, "-F", device).CombinedOutput()
+		// Mount
+		_, err = exec.Command("mkdir", "-p", mountpoint).CombinedOutput()
 		if err != nil {
-			return nil, status.Error(codes.Aborted, fmt.Sprintf("failed to mkfs: %v", err.Error()))
+			return nil, status.Error(codes.Aborted, fmt.Sprintf("failed to mkdir: %v", err.Error()))
 		}
 	} else {
-		if "" != mnt.FsType {
-			if mnt.FsType != curFSType {
-				glog.Errorf("Volume formatted but is incompatible, %v != %v!", mnt.FsType,curFSType )
-				return nil, status.Error(codes.Aborted, "Volume formatted but is incompatible")
-			}
-		}
+		vol.Metadata[KCSIVolumeMode] = "Block"
 	}
 
-	// Mount
-	_, err = exec.Command("mkdir", "-p", mountpoint).CombinedOutput()
-	if err != nil {
-		return nil, status.Error(codes.Aborted, fmt.Sprintf("failed to mkdir: %v", err.Error()))
-	}
-
-	err = mountDeviceAndUpdateAttachment(device, mountpoint, KStagingTargetPath, mountFlags, needUpdateAtc, attachment)
+	err = mountDeviceAndUpdateAttachment(device, mountpoint, KStagingTargetPath, mountFlags, needUpdateAtc, attachment, block)
 	if err != nil {
 		return nil, err
 	}
@@ -312,15 +350,23 @@ func (p *Plugin) NodeUnstageVolume(
 		return nil, status.Error(codes.InvalidArgument, "Volume_id/staging_target_path must be specified")
 	}
 
-	// Umount
-	err := connector.Umount(req.StagingTargetPath)
+	vol, attachment, err := getVolumeAndAttachmentByVolumeId(req.VolumeId)
 	if err != nil {
 		return nil, err
 	}
 
-	vol, attachment, err := getVolumeAndAttachmentByVolumeId(req.VolumeId)
-	if err != nil {
-		return nil, err
+	if KCSIFilesystem == vol.Metadata[KCSIVolumeMode] {
+		err = connector.Umount(req.StagingTargetPath)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	if KCSIBlock == vol.Metadata[KCSIVolumeMode] {
+		_, err = exec.Command("rm", "-rf", req.StagingTargetPath).CombinedOutput()
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	err = delTargetPathInAttachment(attachment, KStagingTargetPath, req.StagingTargetPath)
@@ -366,31 +412,40 @@ func (p *Plugin) NodePublishVolume(
 
 	device := req.StagingTargetPath
 	mountpoint := req.TargetPath
-	needUpdateAtc := false
 
 	// Check if it is: "Volume published but is incompatible"
 	mnt := req.VolumeCapability.GetMount()
-	mountFlags := append(mnt.MountFlags, "bind")
-	if req.Readonly {
-		mountFlags = append(mountFlags, "ro")
+	block := req.VolumeCapability.GetBlock()
+	glog.V(5).Infof("VolumeCapability Mount=%+v, Block=%+v\n", mnt, block)
+
+	if (nil != mnt) && (nil != block) {
+		return nil, status.Error(codes.InvalidArgument, "volumeMode cannot be both Block and Filesystem")
 	}
 
-	_, err = exec.Command("findmnt", device, mountpoint).CombinedOutput()
-	glog.V(5).Infof("findmnt err: %v \n", err)
-
-	if nil == err {
-		if len(mountFlags) > 0 {
-			_, err := exec.Command("findmnt", "-o", strings.Join(mountFlags, ","), device, mountpoint).CombinedOutput()
-			if nil != err {
-				return nil, status.Error(codes.Aborted, "Volume published but is incompatible")
-			}
+	var mountFlags []string
+	if nil == block {
+		mountFlags = append(mnt.MountFlags, "bind")
+		if req.Readonly {
+			mountFlags = append(mountFlags, "ro")
 		}
 
-		return &csi.NodePublishVolumeResponse{}, nil
+		_, err = exec.Command("findmnt", device, mountpoint).CombinedOutput()
+		glog.V(5).Infof("findmnt err: %v \n", err)
+
+		if nil == err {
+			if len(mountFlags) > 0 {
+				_, err := exec.Command("findmnt", "-o", strings.Join(mountFlags, ","), device, mountpoint).CombinedOutput()
+				if nil != err {
+					return nil, status.Error(codes.Aborted, "Volume published but is incompatible")
+				}
+			}
+
+			return &csi.NodePublishVolumeResponse{}, nil
+		}
+
 	}
 
-	// Mount
-	err = mountDeviceAndUpdateAttachment(device, mountpoint, KTargetPath, mountFlags, needUpdateAtc, attachment)
+	err = mountDeviceAndUpdateAttachment(device, mountpoint, KTargetPath, mountFlags, false, attachment, block)
 	if err != nil {
 		return nil, err
 	}
@@ -412,15 +467,23 @@ func (p *Plugin) NodeUnpublishVolume(
 		return nil, status.Error(codes.InvalidArgument, "Volume_id/target_path must be specified")
 	}
 
-	// Umount
-	err := connector.Umount(req.TargetPath)
+	vol, attachment, err := getVolumeAndAttachmentByVolumeId(req.VolumeId)
 	if err != nil {
 		return nil, err
 	}
 
-	_, attachment, err := getVolumeAndAttachmentByVolumeId(req.VolumeId)
-	if err != nil {
-		return nil, err
+	if KCSIFilesystem == vol.Metadata[KCSIVolumeMode] {
+		err := connector.Umount(req.TargetPath)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	if KCSIBlock == vol.Metadata[KCSIVolumeMode] {
+		_, err = exec.Command("rm", "-rf", req.TargetPath).CombinedOutput()
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	err = delTargetPathInAttachment(attachment, KTargetPath, req.TargetPath)
@@ -525,7 +588,7 @@ func (p *Plugin) NodeGetCapabilities(
 	}, nil
 }
 
-// NodeGetVolumeStats
+// NodeGetVolumeStats implementation
 func (p *Plugin) NodeGetVolumeStats(
 	ctx context.Context,
 	req *csi.NodeGetVolumeStatsRequest) (
